add settings UI on "/"
