+++ src/routes/[id]/+page.ts
@@ -1,0 +1,5 @@
+import type { PageLoad } from './$types';
+
+export const load = (async ({ params }) => {
+	return { id: params.id };
+}) satisfies PageLoad;
